Fix various warnings in gio/gdatainputstream.c
In file included from glib/glibconfig.h:9,
                 from glib/gtypes.h:32,
                 from glib/galloca.h:32,
                 from glib/glib.h:30,
                 from gobject/gbinding.h:28,
                 from glib/glib-object.h:23,
                 from gio/gioenums.h:28,
                 from gio/giotypes.h:28,
                 from gio/ginputstream.h:28,
                 from gio/gfilterinputstream.h:28,
                 from gio/gbufferedinputstream.h:28,
                 from gio/gdatainputstream.h:28,
                 from gio/gdatainputstream.c:24:
gio/gdatainputstream.c: In function ‘read_data’:
gio/gdatainputstream.c:313:23: error: comparison of integer expressions of different signedness: ‘gssize’ {aka ‘long int’} and ‘gsize’ {aka ‘long unsigned int’} [-Werror=sign-compare]
   g_warn_if_fail (res == size);
                       ^~
glib/gmacros.h:455:25: note: in definition of macro ‘G_LIKELY’
 #define G_LIKELY(expr) (expr)
                         ^~~~
gio/gdatainputstream.c:313:3: note: in expansion of macro ‘g_warn_if_fail’
   g_warn_if_fail (res == size);
   ^~~~~~~~~~~~~~
gio/gdatainputstream.c: In function ‘scan_for_newline’:
gio/gdatainputstream.c:654:40: error: comparison of integer expressions of different signedness: ‘int’ and ‘gsize’ {aka ‘long unsigned int’} [-Werror=sign-compare]
   for (i = 0; checked < available && i < peeked; i++)
                                        ^
gio/gdatainputstream.c: In function ‘scan_for_chars’:
gio/gdatainputstream.c:879:40: error: comparison of integer expressions of different signedness: ‘int’ and ‘gsize’ {aka ‘long unsigned int’} [-Werror=sign-compare]
   for (i = 0; checked < available && i < peeked; i++)
                                        ^

+++ gio/gdatainputstream.c
@@ -310,7 +310,7 @@ read_data (GDataInputStream  *stream,
   res = g_input_stream_read (G_INPUT_STREAM (stream),
 			     buffer, size,
 			     NULL, NULL);
-  g_warn_if_fail (res == size);
+  g_warn_if_fail (res == (gssize) size);
   return TRUE;
 }
 
@@ -631,7 +631,7 @@ scan_for_newline (GDataInputStream *stream,
   GDataInputStreamPrivate *priv;
   const char *buffer;
   gsize start, end, peeked;
-  int i;
+  gsize i;
   gssize found_pos;
   int newline_len;
   gsize available, checked;
@@ -861,7 +861,7 @@ scan_for_chars (GDataInputStream *stream,
   GBufferedInputStream *bstream;
   const char *buffer;
   gsize start, end, peeked;
-  int i;
+  gsize i;
   gsize available, checked;
   const char *stop_char;
   const char *stop_end;
